% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_country_code.R
\name{as_country_code}
\alias{as_country_code}
\title{Convert standardized country names to country codes}
\usage{
as_country_code(x, to = c("iso3c", "ar5", "cowc", "cown", "eu28", "fao",
  "fips105", "icao", "imf", "ioc", "iso2c", "iso3n", "un", "wb", "eurostat",
  "wb_api2c", "wb_api3c", "p4_scode", "p4_ccode", "wvs"), language = c("en",
  "ar", "de", "es", "fr", "ru", "zh"), factor = FALSE, ...)
}
\arguments{
\item{x}{A vector of country names}

\item{to}{Code format to which to convert. Defaults to \code{"iso3c"}.}

\item{language}{Language of country names supplied. Defaults to \code{"en"}.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}

\item{...}{Other arguments passed to \code{\link[countrycode]{countrycode}}.}
}
\value{
A vector of country codes.
}
\description{
\code{as_country_code} converts a vector of standardized country names to
country codes
}
\details{
Here's a placeholder of a fuller description
}
