% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_country_name.R
\name{as_country_name}
\alias{as_country_name}
\title{Convert standardized country codes to country names}
\usage{
as_country_name(x, from = c("iso3c", "cowc", "cown", "eurostat", "fao",
  "fips105", "imf", "ioc", "iso2c", "iso3n", "p4_ccode", "p4_scode", "un", "wb",
  "wb_api2c", "wb_api3c", "wvs"), language = c("en", "ar", "de", "es", "fr",
  "ru", "zh"), factor = FALSE, ...)
}
\arguments{
\item{x}{A vector of country codes}

\item{from}{Code format from which to convert. Defaults to \code{"iso3c"}.}

\item{language}{Language of country names desired. Defaults to \code{"en"}.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}

\item{...}{Other arguments passed to \code{\link[countrycode]{countrycode}}.}
}
\value{
A vector of country names.
}
\description{
\code{as_country_name} converts a vector of standardized country codes to
country names.
}
\details{
Here's a placeholder of a fuller description
}
