% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{new_vault}
\alias{new_vault}
\title{Initialize a password vault with a given filename}
\usage{
new_vault(filename)
}
\arguments{
\item{filename}{The external file where to store the credentials.
\code{filename} can either be an existing file or a new one. External
files must be in the \code{.rds} file format.}
}
\value{
A \code{passport} vault object.
}
\description{
Initialize a password vault with a given filename
}
\examples{
temp <- tempfile()
new_vault(filename = temp)
}
