% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/passport.R
\docType{package}
\name{passport}
\alias{passport}
\alias{passport-package}
\title{passport package - Manage Access Credentials with R}
\description{
passport is an utility that lets you manage username and password
in order to provide full access to external resources without exposing
credential within the code.
See the full \href{https://github.com/msaltieri/passport}{README} on GitHub.
}
