% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_add}
\alias{vault_add}
\title{Add new credentials to the vault}
\usage{
vault_add(vault, value)
}
\arguments{
\item{vault}{A \code{passport} vault object.}

\item{value}{A list with the new credential to add. \code{vault} must
contain the unique field `service` and the optional fields `host`, `port`,
`username` and `password`.}
}
\description{
Add new credentials to the vault
}
\keyword{internal}
