% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_delete}
\alias{vault_delete}
\title{Delete the entire vault}
\usage{
vault_delete(vault)
}
\arguments{
\item{vault}{A \code{passport} vault object.}
}
\description{
Delete the entire vault
}
\keyword{internal}
