% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_filename}
\alias{vault_filename}
\title{Get the vault filename}
\usage{
vault_filename(vault)
}
\arguments{
\item{vault}{A \code{passport} vault object.}
}
\description{
Get the vault filename
}
\keyword{internal}
