% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_get}
\alias{vault_get}
\title{Get the credentials for a single service}
\usage{
vault_get(vault, service)
}
\arguments{
\item{vault}{A \code{passport} vault object.}

\item{service}{A service to retrieve credentials for}
}
\description{
Get the credentials for a single service
}
\keyword{internal}
