% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_getall}
\alias{vault_getall}
\title{Get all the credentials stored in the vault}
\usage{
vault_getall(vault)
}
\arguments{
\item{vault}{A \code{passport} vault object.}
}
\description{
Get all the credentials stored in the vault
}
\keyword{internal}
