% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_rm}
\alias{vault_rm}
\title{Remove the credentials for a single service}
\usage{
vault_rm(vault, service)
}
\arguments{
\item{vault}{A \code{passport} vault object.}

\item{service}{A service to delete credentials to}
}
\description{
Remove the credentials for a single service
}
\keyword{internal}
