% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_rmall}
\alias{vault_rmall}
\title{Remove all the credentials from the vault}
\usage{
vault_rmall(vault)
}
\arguments{
\item{vault}{A \code{passport} vault object.}
}
\description{
Remove all the credentials from the vault
}
\keyword{internal}
