% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vault.R
\name{vault_services}
\alias{vault_services}
\title{Get the list of services in the vault}
\usage{
vault_services(vault)
}
\arguments{
\item{vault}{A \code{passport} vault object.}
}
\description{
Get the list of services in the vault
}
\keyword{internal}
