% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_country_code.R
\name{as_country_code}
\alias{as_country_code}
\title{Convert standardized country names to country codes}
\usage{
as_country_code(x, from, to = "iso2c", factor = is.factor(x))
}
\arguments{
\item{x}{A character, factor, or numeric vector of country names or codes}

\item{from}{Format from which to convert. See Details for more options.}

\item{to}{Code format to which to convert. Defaults to \code{"iso2c"}; see
Details for more options.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}
}
\value{
A vector of country codes.
}
\description{
\code{as_country_code} converts a vector of standardized country names to
country codes
}
\details{
Here's a placeholder of a fuller description
}
