% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_country_name.R
\name{as_country_name}
\alias{as_country_name}
\title{Convert standardized country codes to country names}
\usage{
as_country_name(x, to = "en", from = "iso2c", short = TRUE,
  variant = FALSE, factor = is.factor(x))
}
\arguments{
\item{x}{A character, factor, or numeric vector of country codes}

\item{to}{Language code of country names desired. Defaults to \code{"en"};
see Details for more options.}

\item{from}{Code format from which to convert. Defaults to \code{"iso2c"};
see Details for more options.}

\item{short}{Whether to use short alternative name when available. Can be
length 1 or the same length as \code{x}.}

\item{variant}{Whether to use variant alternative name when available. Can
be length 1 or the same length as \code{x}.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}
}
\value{
A character or factor vector of country names.
}
\description{
\code{as_country_name} converts a vector of standardized country codes to
country names.
}
\details{
Here's a placeholder of a fuller description
}
