% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_country.R
\name{parse_country}
\alias{parse_country}
\title{Parse country names to standardized form}
\usage{
parse_country(x, to = c("iso3c", "ar5", "continent", "cowc", "cown", "eu28",
  "eurocontrol_pru", "eurocontrol_statfor", "fao", "fips105", "icao",
  "icao_region", "imf", "ioc", "iso2c", "iso3n", "region", "un", "wb",
  "country.name.ar", "country.name.de", "country.name.en", "country.name.es",
  "country.name.fr", "country.name.ru", "country.name.zh", "eurostat",
  "wb_api2c", "wb_api3c", "p4_scode", "p4_ccode", "wvs"), language = c("en",
  "de"), factor = FALSE, ...)
}
\arguments{
\item{x}{A character vector of country names.}

\item{to}{Output format, defaults to \code{"iso3c"}.}

\item{language}{Language from which to parse country names. Currently
accepts \code{"en"} (default) and \code{"de"}.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}

\item{...}{Other arguments passed to \code{\link[countrycode]{countrycode}}.}
}
\value{
A character vector of ISO 3-character country codes, unless
\code{factor = TRUE}. Warns of any parsing failure.
}
\description{
\code{parse_country} parses country names to the standardized ISO
3-character code.
}
\details{
Here's a placeholder of a fuller description
}
