% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_country.R
\name{parse_country}
\alias{parse_country}
\title{Parse country names to standardized form}
\usage{
parse_country(x, to = "iso2c", language = c("en", "de"), short = TRUE,
  variant = FALSE, factor = is.factor(x))
}
\arguments{
\item{x}{A character or factor vector of country names to standardize}

\item{to}{Format to which to convert. Defaults to \code{"iso2c"}; see Details
for more options.}

\item{language}{Language from which to parse country names. Currently
accepts \code{"en"} (default) and \code{"de"}.}

\item{short}{Whether to use short alternative name when available. Can be
length 1 or the same length as \code{x}.}

\item{variant}{Whether to use variant alternative name when available. Can
be length 1 or the same length as \code{x}.}

\item{factor}{If \code{TRUE}, returns factor instead of character vector.}
}
\value{
A character vector of ISO 3-character country codes, unless
\code{factor = TRUE}. Warns of any parsing failure.
}
\description{
\code{parse_country} parses country names to the standardized ISO
3-character code.
}
\details{
Here's a placeholder of a fuller description
}
